app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.7.1
pydantic-settings==2.2.1
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
psycopg2-binary==2.9.9
python-dotenv
pydantic-settings
email-validator
numpy>=1.18.1
decorator>=4.3.0
celery>=5.2.3
pytest
pillow
python-multipart==0.0.6
